requirements.txt
chromadb
sentence-transformers
pymupdf
pdfplumber
pillow
langchain
langchain-google-genai
langgraph
spacy
yfinance
requests
beautifulsoup4
pydantic
fastapi
uvicorn
apscheduler
psycopg2-binary
ragas
deepeval
python-dotenv
google-genai
numpy
sec-edgar-downloader
finbert-embedding

